% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\docType{data}
\name{chimera}
\alias{chimera}
\title{Chimera data}
\format{
A matrix with 8405 genes (rows) and 78 samples (columns).
}
\source{
Sabina Chiaretti, Xiaochun Li, Robert Gentleman, Antonella Vitale, Marco Vignetti, Franco Mandelli, Jerome Ritz, and Robin Foa Gene expression profile of adult T-cell acute lymphocytic leukemia identifies distinct subsets of patients with different response to therapy and survival. Blood, 1 April 2004, Vol. 103, No. 7.
}
\usage{
chimera
}
\description{
A matrix containing the expression values of 3405 genes deriving from Affimetrix single
channel technology, consisting of 41 observations from one experimental condition
(absence of BCR/ABL gene arrangment, class 1), and 37 observations from another
experimental condition (presence of BCR/ABL gene arrangment, class 2).
}
\examples{
data(chimera)
}
\keyword{datasets}
