% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{easyLookDys}
\alias{easyLookDys}
\title{Dysregulation summary}
\usage{
easyLookDys(resObj, digits = 4)
}
\arguments{
\item{resObj}{The output of simPATHy function (simPATHy class object).}

\item{digits}{Integer indicating the number of decimal places to be used.}
}
\value{
Nice formatted output of simPATHy dysregulation

Nicely formatted output of simPATHy dysregulation.
}
\description{
Summary of the result for a quick look of simPATHy function.
}

