% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_function.R
\name{easyLookShiny}
\alias{easyLookShiny}
\title{Visual dysregulation summary}
\usage{
easyLookShiny(resObj, graph, heightGraph = NULL, heightMatrix = NULL)
}
\arguments{
\item{resObj}{The output of simPATHy function}

\item{graph}{The graphNEL object given to the simPATHy function to obtain resObj.}

\item{heightGraph, heightMatrix}{The height of the graph and correlation matrix plots in pixels. Must be a number, which will be coerced to a string and have 'px' append.}
}
\value{
Interactive plots for exploring the output of simPATHy.
}
\description{
A Shiny application for visual summary of dysregulation.
}
\seealso{
\code{\link[simPATHy]{simPATHy}}, \code{\link[simPATHy]{plotGraphNELD3}}, \code{\link[simPATHy]{plotCorGraph}}, \code{\link[simPATHy]{easyLookDys}}
}

