% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{generatePath}
\alias{generatePath}
\title{Find one path in a graph}
\usage{
generatePath(graph, from = NULL, to = NULL)
}
\arguments{
\item{graph}{A directed or undirected graph represented as a graphNEL object.}

\item{from, to}{The nodes (character node id) giving the first and the last nodes of the path to be calculated. If \code{NULL} then the \code{from} and \code{to} nodes are randomly choosen.}
}
\value{
A list of edges in edgesList format (see \pkg{gRbase}).
}
\description{
Find one shortest path in the graph between two given nodes.
}
\seealso{
\code{\link[igraph]{get.all.shortest.paths}}
}
