% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_function.R
\name{getPathShiny}
\alias{getPathShiny}
\title{Choose a path in a graph from an interactive shiny app}
\usage{
getPathShiny(graph)
}
\arguments{
\item{graph}{A graphNEL object.}
}
\value{
Selected path with the right format for simPATHy function.
}
\description{
Choose a path in a graph from an interactive shiny app with the rigth format for simPATHy function.
}
\examples{
 if(require(gRbase)){
  graph <- gRbase::dag(~c:a, ~c:b, ~d:c, ~e:d)

  # Launch the interactive plot
  # path <- getPathShiny(graph)
 }
}
\seealso{
\code{\link[simPATHy]{simPATHy}}
}
