% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphNELD3.R
\name{graphNELD3-shiny}
\alias{graphNELD3-shiny}
\alias{graphNELD3Output}
\alias{renderGraphNELD3}
\title{Shiny bindings for plotGraphNELD3}
\usage{
graphNELD3Output(outputId, width = "100\%", height = "400px")

renderGraphNELD3(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{Output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a graphNELD3}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using plotGraphNELD3 within Shiny
applications and interactive Rmd documents.
}

