% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{SMLEdecomposable}
\alias{SMLEdecomposable}
\title{Local Maximum Likelihood Estimation}
\usage{
SMLEdecomposable(S, graph)
}
\arguments{
\item{S}{a covariance matrix.}

\item{graph}{a decomposable graph represented as a graphNEL object.}
}
\value{
The MLE of a covarince matrix.
}
\description{
Compute a maximum likelihood estimate of a covariance matrix in a decomposable Gaussian graphical model.
}
\references{
Lauritzen, S. L. (1996). Graphical Models. Clarendon Press, Oxford.
}
